backend/requirements/base.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-axes==6.5.0
argon2-cffi==23.1.0
django-filter==24.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-spectacular==0.27.2
Pillow==10.4.0
